- Improved handling of missing or incompatible protocols, including a case that was previously overlooked (Graham) - Fix bug in S3C80 translation that was causing overflow messages in certain circumstances (Graham)
git-svn-id: https://svn.code.sf.net/p/controlremote/code/trunk@966 bfc585dc-5255-483b-a5ca-a9bf5897b33d

+++ km/com/hifiremote/jp1/S3C80Processor.java
@@ -127,16 +127,16 @@ public class S3C80Processor
         int second = data[ ++i ] & 0xFF;
         if ( codeType == CodeType.NEW && second == 0xFF )
         {
-          data[ i ] = ( byte )0x80;
+          data[ i ] = ( short )0x80;
         }
         else if ( codeType == CodeType.OLD && second == 0x80 )
         {
-          data[ i ] = ( byte )0xFF;
+          data[ i ] = ( short )0xFF;
         }
         else if ( second == 0x01 )
         {
           int third = data[ ++i ] & 0xFF;
-          data[ i ] = ( byte )adjust( third, codeType );
+          data[ i ] = ( short )adjust( third, codeType );
         }
       }
     }
